rename container to collection
