+++ crates/hir/src/collection/uninit.rs
