+++ crates/parser/src/lib.rs
@@ -1,5 +1,5 @@
+pub mod collections;
 pub mod conditionals;
-pub mod containers;
 pub mod cst;
 pub mod delimited;
 pub mod errors;
